rebased fixed my changes to work with static columns
