+++ packages/storybook/src/ui/organisms/questionnaire-table/questionnaire-table.stories.tsx
@@ -33,6 +33,7 @@ export const HorizontalTableAdminActions: Story = () => (
 export const VerticalTableWithPagination: Story = () => (
   <QuestionnaireTable
     tableOptions={{ enablePagination: true, serverSidePagination: true, pageSize: 1, perPageOptions: [1, 10, 20] }}
+    staticColumns={2}
     fetchData={(options) =>
       new Promise((res) =>
         res({
